Initial Setup
When a page is vulnerable to enumeration, it is because there is some
aspect of its behavior or content based on whether or not certain
input data is valid, in this case the username.

LoginFunctionTest#testForEnumeration and the two Registraion jmh tests
demonstrate Terracotta Bank's vulnerability enumeration vulnerabilies.

In this commit, you can see the changes that were made in preparation
for the Enumeration demos.