Unify Messaging
One way to combat enumeration is to make sure that all messaging is
the same, regardless of the server-side outcome of a user operation.

In the case of a failed login, state "The username or password is
invalid". In the case of a failed registration due to a username
conflict, state "You will receive instructions soon to the email
address provided".