Mitigate Bruteforce with Soft Lockout
One can mitigate brute force by tracking the number of failed login
attempts against each individual account. This can be done in a cache
so that invalid usernames can also be tracked, preventing enumeration
attacks via the lockout mechanism.

Check out the next commit for details on taking care of the DoS
vulnerability that is opened up because of this approach.