Mitigate Brute Force with IP Soft Lockout
When an attacker wants to get passed the account-level throttling,
it is trivial to simply invert the attack and try the same password
across many accounts. This can be addressed with IP throttling at
the networking level, though this commit shows what similar
application logic might look like.

Also in this commit is a secure way to discover the ip address for
a client.