Mitigating Plaintext Using security-constraint
Using the Servlet Spec, we can force our application to use https
when http traffic comes in. This is done with the security-constraint
attribute.

Now, it is possible that this concern is taken care of in a different
portion of your infrastructure, like a load balancer; however, this
can still be useful for local testing, even in those circumstances.