Soft Lockout Cache Clear
In some cases, we can preemptively clear the soft lockout cache so
that tests do not affect one another. For example, because several
brute force tests hit the login endpoint with many consecutive login
failures, they jointly trigger the IP lockout. If we inject a
successful login between each of these tests, it will reset the
counter for that lockout.

Note that IP lockout tests still need to be run separately, though.