Password Complexity - Special Characters, Length
Passwords are hashed in the end, so what the user puts in their
password is typically harmless to the underlying system storing it.
Because they are hashed, the contents become a predictable
character set and consistent length.

So, users should be allowed a wider range of special characters than
many sites allow and longer passwords than allowed as well.