Password Complexity - Nbvcxz
Password complexity is best thought of in terms of entropy, and the
nbvcxz library performs a rough calculation of entropy based on the
contents of passwords.

Using a library like nbvcxz allows us an alternative to entropy
surrogates like LUNS and also allows us to do a bit of education of
our users, helping them to select better passwords.