Initial Setup - Password Upgrades
To demonstrate upgrading passwords, we'll use the Spring Security
password encoder API. As such, the basic user portion of the
application is converted over to use Spring Security with noop
password encoding.

You can see that the password tests still fail here since, because
there is no hashing or salting on the passwords, it is still
vulnerable to the same exploitations demonstrated in previous modules.