Upgrading Plaintext Passwords - Active Migration
We can use Spring Security's password encoder API to assist us in
performing an active migration of our passwords to use a modern
hashing algorithm.

The easiest case is when our passwords are in plaintext and we can
simple do a batch run using Spring Security's default delegating
password encoder.