Change Password - Transactional Authority
In order to change an element of an authentication mechanism, we need
to require at least the same level of authority that it grants.

For example, if we want to change a password, we need to obtain the
old password or we need to perform some stronger kind of authorization
than that.

If we want to change a second factor, then we need to first require
the second factor or we need to perform some stronger kind of
authorization than that.