Transactional Authority - Transaction Keys
In general, a key should be associated with the authorities that it
is granted and on whose behalf. So, when the user says he forgot his
password, we generate a key and tie it to him as a user and to the
permission the app grants to this key that it can change the password.

Then, when we go to the change password request, we check the provided
key for the permissions. If the key is still in the cache, and the
permissions match, we show the page. On submission, we check again.
Once the password is updated, then we invalidate the key and they are
now free to login with their new password.