Authentication Logging - Authentication Events
We want to log whenever authentication events pass or fail, and why.
We also want to log whenever authorization events pass or fail, and
whenever sensitive data access requests pass or fail, and why.

Note that in this commit, all of these are configured to be on, but it
would likely be necessary to add additional domain-specific logic for
authorization-success event logging to be informative.